Maj_projet_fil_rouge(exo3_ET_4)

+++ learning_java/src/main/java/fr/hetic/Exo3/Addition.java
@@ -2,7 +2,8 @@ package fr.hetic.Exo3;
 
 public class Addition implements Operation {
     @Override
-    public double execute(double num1, double num2) {
+    public int execute(int num1, int num2) {
         return num1 + num2;
     }
 }
+
